. dbusport.cpp: fixed a bug that lpbk message could arrive after the dbusport is stopped, because it is passed to the peer via RescheduleTask insteamd RescheduleTaskMainLoop. and only the latter can guarantee the message's lifecycle is contained within dbusport's lifecycle. . dbusprxy.cpp: fixed a bug that the irp with dbus message was released after the dbus was down and caused segment fault. The fix is to release the dbus message in the CompleteConnReq where it is assured to be safe to handle the dbus messages.
. ifhelper.h, iftasks.cpp: fixed a regression bug when changed to use
CTaskWrapper as the callback of the completion function, that the transferring
of response parameters was missed to change its destination from the inner task
to CTaskWrapper. And added 'return code' as the input argument to
TransferParams, so that at least the true error code can be passed to the
completion function in the default condition.

. ifhelper.h, rpcif.cpp: Added a specifical version of AddSeqTaskTempl for
CRpcServices with extra check of the CRpcServices' curren state.

. proxy.h, rpcif.cpp, reqfwdr.cpp: rewrote RunManagedTask with additional check
of the CRpcServices' state to determine wheter to proceed or return with error.
It is an important way to control the lifecycle of the task to run.

. stream.cpp, stmrelay.cpp, stmrelay.h, streammh.cpp: made a change to the way
to complete the family of CIfStartUxSockStmTask, to make sure StartTicking won't
be added to the m_pRootTaskGroup when the StartEx's tasks are still in the
m_pRootTaskGroup.