. rpcroute.cpp: fixed some bugs which prevent CRpcRouterBridge's OnPreStopLongWait to work properly, especially missed a callback from OnClose to OnRmtSvrEvent serving as notification to move forword CRpcRouterBridge's Stop process.
. bridge.cpp: added call to m_pHsTicker's RemoveTimer when the handshake
completes before the timeout. There is a memory leak if the rpcrouter shutdown
happens before the timer was due, because the old code assumed another timer-id
of CIfDeferCallTaskEx as the one taking effect, and removed it via DisableTimer
on complete, but the correct one is still working and holds a reference to the
CIfDeferCallTaskEx even after the CIoManager is completely stopped.