. clsids.h, dbuspdo.cpp, dbusport.cpp, dbusport.h, dllexp.cpp, namespc.h: added CDBusLoopbackPdo2, which unlike CDBusLoopbackPdo, transfer events as well as request/response, between server and client, it will broadcast event only to clients that have registered matches with it.
. testypes/maincli.cpp, iomgr.cpp, frmwrk.h, ifstat.cpp, ifstat.h, rpcbase.i,
gencpp.cpp, rpcf.sip, rtauth.json, synccfg.py.in, rpcroute.cpp: added
propBuiltinRt property to label apps with builtin-rt. And made
CRpcRouterBridgeAuthImpl and CRpcReqForwarderProxyAuthImpl to communicate via
CDBusLoopbackPdo2 all the time. And enable CRpcReqForwardProxyImpl to
communicate with user's server with CDBusLoopbackPdo2 when propBuiltinRt exists
and propRouterRole has routerbridge, which fixed the bug of many timeout
FetchData when system is stressed with huge surge of incoming connections.